LegionBoard: fix bug with course being "null"
When upgrading LegionBoard Heart from 0.1.x to 0.2+, all existing changes get the course "null". Mistakenly, I thought the JSONObject would contain a "real" null, but the null is given as a string.

The following IOException is fixed with this commit:
java.io.IOException: Change references a course but courses are empty.
	at me.vertretungsplan.parser.LegionBoardParser.getSubstitution(LegionBoardParser.java:201)
	at me.vertretungsplan.parser.LegionBoardParser.parseLegionBoard(LegionBoardParser.java:160)
	at me.vertretungsplan.parser.LegionBoardParser.getSubstitutionSchedule(LegionBoardParser.java:69)